update: ユーザーの情報 コンテキストメニューのリファクタリング
